+++ servidor/src/main/java/br/com/empresa/rh/service/folha/IEvento.java
@@ -16,7 +16,7 @@ public interface IEvento {
 
     Evento getEvento();
 
-    int getReferencia();
+    double getReferencia();
 
     double getValorCalculado();
